% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Kanpsack Dynamic}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{is a data frame reprents the totla items, containing the two colums w and v, w represents thw weight of item and v represnets the value}

\item{W}{is an integer that represnts the maximun weight of the container}
}
\value{
It return the list, which the total value of item in knapsack and their position in the data frame
}
\description{
Fills the bag with  best weights and values
}
\details{
Knapsack
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack problem#0.2F1 knapsack problem}
}
