% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_data.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Mocked dateset}
\format{
A data.frame with 2000 rows and two variables:
\describe{
 \item{w}{Integer,  weight}
 \item{v}{float, value of each object}
}
}
\usage{
knapsack_objects
}
\description{
Mocked dateset
}
\keyword{datasets}
