\name{knapsack_package-package}
\alias{knapsack_package-package}
\alias{knapsack_package}
\docType{package}
\title{
\packageTitle{knapsack_package}
}
\description{
\packageDescription{knapsack_package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{knapsack_package}
\packageIndices{knapsack_package}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{knapsack_package}

Maintainer: \packageMaintainer{knapsack_package}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
