% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_parallel.R
\name{parallel_knapsack_brute_force}
\alias{parallel_knapsack_brute_force}
\title{Kanpsack Bruteforce}
\usage{
parallel_knapsack_brute_force(x, W, parallel = FALSE)
}
\arguments{
\item{x}{is a data frame represents the total items, containing the two columns w and v, w represents thw weight of item and v represnets the value}

\item{W}{is an integer that represents the maximum weight of the container}

\item{parallel}{It is set to FALSE as default, if you set it to TRUE the function will be parallellized.}
}
\value{
It return the list, which the total value of item in knapsack and their position in the data frame
}
\description{
'Parallelize brute force search
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack problem#0.2F1 knapsack problem}
}
