% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_knapsack.R
\name{knapsack_brute_force}
\alias{knapsack_brute_force}
\title{Brute Force Knapsack Problem}
\usage{
knapsack_brute_force(x, W)
}
\arguments{
\item{x}{is a dataframe with two variables v and w}

\item{W}{is the knapsack size}
}
\value{
a list of max_value and chosen items
}
\description{
Brute Force Knapsack Problem
}
\references{
https://en.wikipedia.org/wiki/Knapsack
}
\author{
Zahra Jalilpour, Arslan Shirazi
}
